+++ hal/CHANGELOG.md
@@ -13,6 +13,8 @@
   - RTIC spawn task API will now require `fugit::Duration<_, _, _>` (aliased
   at `atsamd_hal::rtc::Duration`) instead of `embedded_time::Duration`
 - Add Device Service Unit (DSU), Non-volatile Controller (NVM), SmartEEPROM support (#526)
+- Expand Public Key Cryptography Controller (PUKCC) to support modular
+  exponentiation (RSA) (#544)
 ---
 
 Changelog tracking started at v0.13